moving tests.utils to abelian.utils
