+++ abelian/utils.py
@@ -1,10 +1,13 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-import itertools
 import functools
+import itertools
+import random
 import types
+
 import numpy as np
+from sympy import Integer, Float, Rational
 
 
 def mod(a, b):
@@ -183,9 +186,39 @@ def copy_func(f):
     return g
 
 
-argmin = functools.partial(arg, min_or_max = min)
-argmax = functools.partial(arg, min_or_max = max)
+argmin = functools.partial(arg, min_or_max=min)
+argmax = functools.partial(arg, min_or_max=max)
+
+
+def close(a, b):
+    numeric_types = (float, int, complex, Integer, Float, Rational)
+    if isinstance(a, numeric_types) and isinstance(a, numeric_types):
+        return abs(a - b) < 10e-10
+    return sum(abs(i - j) for (i, j) in zip(a, b))
+
+
+def random_zero_heavy(low, high):
+    """
+    Draw a random number, with approx 50% probability of zero.
+    """
+    return random.choice(list(range(low, high)) + [0] * (high - low))
+
+
+def frob_norm(A, B):
+    """
+    Frobenius norm.
+    """
+    return sum(abs(i - j) for (i, j) in zip(A, B))
+
+
+def random_from_list(number, list_to_take_from):
+    """
+    Draw several random values from the same list.
+    """
+    return [random.choice(list_to_take_from) for i in range(number)]
+
 
 if __name__ == "__main__":
     import doctest
-    doctest.testmod(verbose = True)
+
+    doctest.testmod(verbose=True)
